New Floating Action Button.
\n

Removed the old floating action button from the recipes app and created a new one in the general templates.

Altered the navbar to remove the old button and add the new one.

Altered the navbar to no longer open the create recipe modal as that is being moved to its own page.

New floating button has an action to add recipe as well as add ingredients. Add recipe sends you to recipes.create and add ingredients sends you to ingredients.search.