Recipe Creation works!
Added JavaScript functionality to Recipe Creation so all the fields now self expand when there is a change in the previous fields.

Added logic into the CreateRecipe view that actually saves the recipe in the database or pops an error saying that there is already a drink with that name in the database.