Authenticate users on all views -  Hide Create Recipe button - complete
Authenticate users on all views -  User Profile get - complete (already was, but making note of it)
Authenticate users on all views -  Hide Profile nav bar button - complete (already was, but making note of it)
Modified the user profile to use tabs for displaying Recipe Favorites vs. Created Recipes.
Modified navbar to include the fab button in it so that it gets served when logged in.
Added registering models for admin site.